refactor: rename and restructure /ui/header folder
